+++ sources/VS/Device/Hardware/HAL/HAL_win.cpp
@@ -7,7 +7,7 @@
 
 uint HAL_GetTick()
 {
-    return clock();
+    return (uint)clock();
 }
